Expanded macro vars to function on the base system.
Refactored macro var code.

Macro vars now function with /macro show
